apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg[binary]==3.2.5
pydantic-settings==2.7.1
python-jose==3.3.0
passlib==1.7.4
Jinja2==3.1.5
